requirements.txt
streamlit>=1.32.0
pdfplumber>=0.10.3
requests>=2.31.0
python-dotenv>=1.0.1
langchain>=0.1.16
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
